+++ twython/__init__.py
@@ -18,9 +18,9 @@ Questions, comments? ryan@venodesigns.net
 """
 
 __author__ = 'Ryan McGrath <ryan@venodesigns.net>'
-__version__ = '2.10.1'
+__version__ = '3.0.0'
 
-from .twython import Twython
+from .api import Twython
 from .streaming import TwythonStreamer
 from .exceptions import (
     TwythonError, TwythonRateLimitError, TwythonAuthError,