+++ twython/helpers.py
@@ -1,3 +1,13 @@
+# -*- coding: utf-8 -*-
+
+"""
+twython.helpers
+~~~~~~~~~~~~~~~
+
+This module contains functions that are repeatedly used throughout
+the Twython library.
+"""
+
 from .compat import basestring, numeric_types
 
 
@@ -6,7 +16,7 @@ def _transparent_params(_params):
     files = {}
     for k, v in _params.items():
         if hasattr(v, 'read') and callable(v.read):
-            files[k] = v
+            files[k] = v  # pragma: no cover
         elif isinstance(v, bool):
             if v:
                 params[k] = 'true'
@@ -14,6 +24,8 @@ def _transparent_params(_params):
                 params[k] = 'false'
         elif isinstance(v, basestring) or isinstance(v, numeric_types):
             params[k] = v
+        elif isinstance(v, list):
+            params[k] = ','.join(v)
         else:
-            continue
+            continue  # pragma: no cover
     return params, files